Fix dtype mismatch: use bfloat16 for ego_history tensors
Model weights are bfloat16, so ego_history_xyz and ego_history_rot
must also be bfloat16 to avoid 'mat1 and mat2 must have the same dtype' error.

+++ app.py
@@ -1822,10 +1822,11 @@ def run_inference_impl(sample_bundle: Optional[Dict], user_command: str, num_sam
                 # ego_history_xyz shape: (B, n_traj_group, num_history_steps, 3)
                 # ego_history_rot shape: (B, n_traj_group, num_history_steps, 3, 3)
                 # Using zeros as placeholder since we don't have real egomotion history
+                # Must use bfloat16 to match model weights
                 model_inputs = {
                     "tokenized_data": {k: v.to(device) for k, v in inputs.items()},
-                    "ego_history_xyz": torch.zeros(1, 1, 16, 3, device=device, dtype=torch.float32),
-                    "ego_history_rot": torch.zeros(1, 1, 16, 3, 3, device=device, dtype=torch.float32),
+                    "ego_history_xyz": torch.zeros(1, 1, 16, 3, device=device, dtype=torch.bfloat16),
+                    "ego_history_rot": torch.zeros(1, 1, 16, 3, 3, device=device, dtype=torch.bfloat16),
                 }
 
                 # Number of trajectory samples
